refactor: code simplification, merge admin and thread set moderator component into one, reuse on both views
